Bioscope Parser
This is a java program to parse the sentences from bioscope data and write it to a text file without the tags. It could be run through terminal as such:
   javac Bioscope.java 
  java Bioscope <bioscopefile.xml>
It will then write all the sentences into a text file named the same as the input file given, but with .txt extension instead of .xml, in this case bioscopefile.txt.